requirements.txt
openai~=1.58.1
mysql-connector-python~=9.5.0
numpy~=1.26.4
pandas~=2.2.2
statsmodels~=0.14.2
streamlit~=1.37.1